requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data Fetching
yfinance>=0.2.28

# Trading API
alpaca-py>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# GUI Application
streamlit>=1.28.0
plotly>=5.17.0

# Technical Indicators
ta-lib>=0.4.28

# Optional: If ta-lib installation fails, use pandas-ta instead
# pandas-ta>=0.3.14

# Logging and Configuration
python-dotenv>=1.0.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools (Optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
